test: switch to Testcontainers PostgreSQL and rename SearchControllerIntegrationTest
