+++ build.gradle
@@ -108,11 +108,11 @@ dependencies {
     testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
 
     // TestContainers for integration tests
-    testImplementation 'org.testcontainers:testcontainers:1.20.4'
-    testImplementation 'org.testcontainers:junit-jupiter:1.20.4'
-    testImplementation 'org.testcontainers:postgresql:1.20.4'
-    testImplementation 'org.testcontainers:elasticsearch:1.20.4'
-    testImplementation 'org.testcontainers:minio:1.20.4'
+    testImplementation 'org.testcontainers:testcontainers:1.21.3'
+    testImplementation 'org.testcontainers:junit-jupiter:1.21.3'
+    testImplementation 'org.testcontainers:postgresql:1.21.3'
+    testImplementation 'org.testcontainers:elasticsearch:1.21.3'
+    testImplementation 'org.testcontainers:minio:1.21.3'
     testImplementation 'com.redis:testcontainers-redis:2.2.4'
 
     testAnnotationProcessor 'org.projectlombok:lombok'
