+++ src/test/resources/application-test.yml
@@ -1,17 +1,10 @@
+# Test profile overrides.
+# Datasource is injected dynamically by Testcontainers PostgreSQL in IntegrationTestBase.
+# All other shared test properties live in application.properties.
 spring:
-  datasource:
-    url: jdbc:h2:mem:testdb
-    driver-class-name: org.h2.Driver
-    username: sa
-    password:
   jpa:
     hibernate:
       ddl-auto: create-drop
-    database-platform: org.hibernate.dialect.H2Dialect
     defer-datasource-initialization: true
   elasticsearch:
     uris: http://localhost:9200
-  data:
-    redis:
-      host: localhost
-      port: 6379
