+++ src/test/resources/application.properties
@@ -1,68 +1,61 @@
 spring.application.name=demo
 
-# Use H2 in-memory database for tests
+# Default datasource: H2 in-memory for unit tests.
+# Integration tests override this with Testcontainers PostgreSQL via @DynamicPropertySource.
 spring.datasource.url=jdbc:h2:mem:testdb;DB_CLOSE_DELAY=-1
 spring.datasource.driver-class-name=org.h2.Driver
 spring.datasource.username=sa
 spring.datasource.password=
 
-# H2 Console (optional, useful for debugging)
-spring.h2.console.enabled=true
-
-# JPA/Hibernate settings for H2
+# JPA/Hibernate
 spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
 spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.defer-datasource-initialization=true
-
-# Disable data.sql execution in tests
 spring.sql.init.mode=never
 
-# JWT Configuration for tests
+# JWT
 jwt.secret=test-secret-key-for-testing-purposes-must-be-at-least-256-bits-long-for-hs256-algorithm
 jwt.expiration=86400000
+jwt.refresh-token-expiration=604800000
 
-# File Upload Configuration for tests
+# File Upload
 app.upload.dir=${java.io.tmpdir}/proje-pazari-test-uploads
 spring.servlet.multipart.enabled=true
 spring.servlet.multipart.max-file-size=5MB
 spring.servlet.multipart.max-request-size=5MB
 
-# Logging
-logging.level.org.springframework.security=WARN
-# Storage Configuration for tests (use local storage to avoid MinIO dependency)
+# Storage (local to avoid MinIO dependency)
 storage.provider=local
 storage.local.path=${java.io.tmpdir}/proje-pazari-test-uploads
 storage.max-file-size=10MB
 storage.allowed-content-types=image/jpeg,image/png,image/gif,image/webp,application/pdf
 storage.redirect-to-presigned=false
 
-# MinIO Configuration (required even if not used, to prevent placeholder resolution errors)
+# MinIO (required to prevent placeholder resolution errors)
 minio.url=http://localhost:9000
 minio.access-key=minioadmin
 minio.secret-key=minioadmin123
 minio.bucket-name=test-bucket
 
-# Disable Elasticsearch for tests (enabled only in specific tests)
+# Disable Elasticsearch by default (enabled only in specific tests)
 spring.data.elasticsearch.enabled=false
 spring.data.elasticsearch.repositories.enabled=false
 spring.autoconfigure.exclude=org.springframework.boot.autoconfigure.elasticsearch.ElasticsearchRestClientAutoConfiguration,org.springframework.boot.autoconfigure.data.elasticsearch.ElasticsearchDataAutoConfiguration,org.springframework.boot.autoconfigure.data.elasticsearch.ElasticsearchRepositoriesAutoConfiguration,org.springframework.boot.autoconfigure.data.redis.RedisAutoConfiguration,org.springframework.boot.autoconfigure.data.redis.RedisRepositoriesAutoConfiguration
 
-# Mail Configuration for tests
+# Mail
 spring.mail.host=localhost
 spring.mail.port=3025
 spring.mail.username=test
 spring.mail.password=test
 app.email.from=test@example.com
 app.frontend.url=http://localhost:3000
-
-# Disable mail health indicator for tests (JavaMailSender is mocked)
 management.health.mail.enabled=false
 
-# Redis Configuration for tests
+# Redis
 spring.data.redis.host=localhost
 spring.data.redis.port=6379
 spring.data.redis.database=0
 spring.data.redis.password=
 
-# JWT Refresh Token
-jwt.refresh-token-expiration=604800000
+# Logging
+logging.level.org.springframework.security=WARN
